api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
gensim==4.3.0
nltk==3.8.1
supabase==1.0.3
openai==0.27.8
python-dotenv==1.0.0
